Abstract Class for Confidence Intervals

Abstract base class for confidence interval measures. See section Inheriting on how to add a new method.

Details

The aggregator of the wrapped measure is ignored, as the inheriting CI dictates how the point estimate is constructed.

Inheriting

To define a new CI method, inherit from the abstract base class and implement the private method: ci(tbl: data.table, rr: ResampleResult, param_vals: named list()) -> numeric(3) Here, tbl contains the columns loss, row_id and iteration, which are the pointwise loss, the identifier of the observation and the resampling iteration. It should return a vector containing the estimate, lower and upper boundary in that order.

Super class

mlr3::Measure -> MeasureAbstractCi

Public fields

resamplings

(character())
On which resamplings this method can operate.

measure

(Measure)

Methods

Public methods

MeasureAbstractCi$new()

MeasureAbstractCi$aggregate()

MeasureAbstractCi$clone()

Inherited methods

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureAbstractCi$new(
  measure = NULL,
  param_set = ps(),
  packages = character(),
  resamplings,
  label
)

Arguments

measure

(Measure)
The measure for which to calculate a confidence interval. Must have $obs_loss.

param_set

(ParamSet)
Set of hyperparameters.

packages

(character())
Set of required packages. A warning is signaled by the constructor if at least one of the packages is not installed, but loaded (not attached) later on-demand via requireNamespace().

resamplings

(character())
To which resampling classes this measure can be applied.

label

(character(1))
Label for the new instance.

Method aggregate()

Obtain a point estimate, as well as lower and upper CI boundary.

Usage

MeasureAbstractCi$aggregate(rr)

Arguments

rr

(ResampleResult)
The resample result.

Returns

named numeric(3)

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureAbstractCi$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.